+++ source/nanoFramework.System.Net.Http/Properties/AssemblyInfo.cs
@@ -10,6 +10,9 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyProduct("System.Net.Http")]
 [assembly: AssemblyCopyright("Copyright © nanoFramework Contributors 2018")]
 
+//////////////////////////////////////////////////////
+// this assembly does NOT have a native counterpart //
+//////////////////////////////////////////////////////
 
 // Setting ComVisible to false makes the types in this assembly not visible 
 // to COM components.  If you need to access a type in this assembly from