+++ CHANGELOG.md
@@ -1,5 +1,177 @@
 # Change Log
 
+## [v1.0.8-preview.1](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.8-preview.1) (2019-08-30)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.6-preview.33...v1.0.8-preview.1)
+
+**Documentation and other chores:**
+
+- Update 3 NuGet dependencies [\#45](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/45)
+
+## [v1.0.6-preview.33](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.6-preview.33) (2019-07-29)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.6-preview.30...v1.0.6-preview.33)
+
+**Documentation and other chores:**
+
+- Update System.Net dependency to 1.2.0-preview.1 [\#44](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/44)
+
+## [v1.0.6-preview.30](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.6-preview.30) (2019-07-18)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.6-preview.26...v1.0.6-preview.30)
+
+**Documentation and other chores:**
+
+- Update 3 NuGet dependencies [\#43](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/43)
+- Update readme.txt [\#42](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/42)
+
+## [v1.0.6-preview.26](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.6-preview.26) (2019-06-23)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.6-preview.19...v1.0.6-preview.26)
+
+**Documentation and other chores:**
+
+- Update 3 NuGet dependencies [\#41](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/41)
+- Work on native assembly declaration [\#39](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/39) [[Breaking-Change](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/labels/Breaking-Change)]
+
+## [v1.0.6-preview.19](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.6-preview.19) (2019-06-20)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.6-preview.17...v1.0.6-preview.19)
+
+**Documentation and other chores:**
+
+- Update 3 NuGet dependencies [\#38](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/38)
+
+## [v1.0.6-preview.17](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.6-preview.17) (2019-06-14)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.6-preview.14...v1.0.6-preview.17)
+
+## [v1.0.6-preview.14](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.6-preview.14) (2019-06-12)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.6-preview.10...v1.0.6-preview.14)
+
+**Documentation and other chores:**
+
+- Update dependencies [\#35](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/35)
+
+## [v1.0.6-preview.10](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.6-preview.10) (2019-06-12)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.6-preview.6...v1.0.6-preview.10)
+
+**Documentation and other chores:**
+
+- Update 3 NuGet dependencies [\#34](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/34)
+
+## [v1.0.6-preview.6](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.6-preview.6) (2019-06-05)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.6-preview.4...v1.0.6-preview.6)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.System.Net from 1.1.0-preview.1 to 1.1.1-preview.1 [\#32](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/32)
+
+## [v1.0.6-preview.4](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.6-preview.4) (2019-06-05)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.6-preview.2...v1.0.6-preview.4)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.System.Net from 1.0.6-preview-026 to 1.1.0-preview.1 [\#31](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/31)
+
+## [v1.0.6-preview.2](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.6-preview.2) (2019-05-31)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-038...v1.0.6-preview.2)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.System.Net from 1.0.6-preview-024 to 1.0.6-preview-026 [\#29](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/29)
+
+## [v1.0.5-preview-038](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-038) (2019-04-24)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-036...v1.0.5-preview-038)
+
+**Documentation and other chores:**
+
+- Update 2 NuGet dependencies [\#28](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/28)
+
+## [v1.0.5-preview-036](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-036) (2019-04-23)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-031...v1.0.5-preview-036)
+
+**Documentation and other chores:**
+
+- Update 3 NuGet dependencies [\#27](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/27)
+
+## [v1.0.5-preview-031](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-031) (2019-04-07)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-028...v1.0.5-preview-031)
+
+**Documentation and other chores:**
+
+- Update 2 NuGet dependencies [\#26](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/26)
+
+## [v1.0.5-preview-028](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-028) (2019-04-05)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-026...v1.0.5-preview-028)
+
+**Documentation and other chores:**
+
+- Improve documentation comments [\#25](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/25)
+
+## [v1.0.5-preview-026](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-026) (2019-04-05)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-024...v1.0.5-preview-026)
+
+**Implemented enhancements:**
+
+- Add SslProtocol property to allow selecting it [\#24](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/24)
+
+## [v1.0.5-preview-024](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-024) (2019-04-03)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-022...v1.0.5-preview-024)
+
+**Documentation and other chores:**
+
+- Update 2 NuGet dependencies [\#23](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/23)
+
+## [v1.0.5-preview-022](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-022) (2019-03-27)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-020...v1.0.5-preview-022)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.System.Net from 1.0.6-preview-011 to 1.0.6-preview-013 [\#22](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/22)
+
+## [v1.0.5-preview-020](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-020) (2019-03-26)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-017...v1.0.5-preview-020)
+
+## [v1.0.5-preview-017](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-017) (2019-03-26)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-009...v1.0.5-preview-017)
+
+## [v1.0.5-preview-009](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-009) (2019-03-25)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-007...v1.0.5-preview-009)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.System.Net from 1.0.6-preview-009 to 1.0.6-preview-011 [\#19](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/19)
+
+## [v1.0.5-preview-007](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-007) (2019-03-15)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-004...v1.0.5-preview-007)
+
+**Fixed bugs:**
+
+- Fix getting local address [\#18](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/18)
+
+## [v1.0.5-preview-004](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-004) (2019-03-11)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-002...v1.0.5-preview-004)
+
+**Documentation and other chores:**
+
+- Bumps nanoFramework.System.Net from 1.0.6-preview-005 to 1.0.6-preview-009 [\#17](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/17)
+
+## [v1.0.5-preview-002](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-002) (2019-03-11)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.5-preview-001...v1.0.5-preview-002)
+
+**Documentation and other chores:**
+
+- Update 2 NuGet dependencies [\#16](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/16)
+
+## [v1.0.5-preview-001](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.5-preview-001) (2019-03-11)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.4-preview-009...v1.0.5-preview-001)
+
+**Documentation and other chores:**
+
+- Update 2 NuGet dependencies [\#15](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/15)
+
+## [v1.0.4-preview-009](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.4-preview-009) (2019-03-10)
+[Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.4...v1.0.4-preview-009)
+
+**Documentation and other chores:**
+
+- Update 2 NuGet dependencies [\#14](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/pull/14)
+
 ## [v1.0.4](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/tree/v1.0.4) (2019-01-22)
 [Full Changelog](https://github.com/nanoframework/lib-nanoFramework.System.Net.Http/compare/v1.0.4-preview-003...v1.0.4)